requirements.txt
absl-py==2.1.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.4
flatbuffers==24.3.25
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
idna==3.10
imageio==2.36.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.13.3
networkx==3.4.2
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-image==0.24.0
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.2
streamlit==1.40.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow
termcolor==2.5.0
tifffile==2024.9.20
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
Werkzeug==3.1.3
wrapt==1.16.0